Add a bottomNavigationBar widget to BasePage.
Add app_settings.dart.

The bottomNavigationBar height is calculated using a scale factor applied to the appBar height. In order to be able to extract the appBar height from the BuildContext, a Builder class is used which delegates its build to a callback. This opens the parent widget's BuildContext for interrogating and getting the appBar height.