Add setState method to initState.
Without setState the post frame callback changes homeButton, settingsButton and filesButton in _ButtonArrayState state, but it doesn't trigger a rebuild. This accounts for the missing home button when transitioning to the next page in the previous commit.

Adding setState triggers a rebuild, although the build begins at the same time as the page transition starts. This means that the button is moving before the page transition is complete.