Fix problem with Button class.
The previous commit showed the button decoration associated with the Ink class moving across the screen in sync with the button icon but behind eveything that was displayed.

This commit uses keys and the CircularAvatar class in place of the Ink class to solve the above problem. At this point it is not know which solves the issue.