Increase number of SettingsTile instances.
Example settings page now scrolls up and down, on and off the page.

Note. There is an issue with BasePage. Currently, when a transition occurs the page zooms into existence, but the new buttons which begin their animation off screen also zoom into existence. This defeats the object of animating the transition.