Save. Not working.
NotificationNotifier has been changed to accept a null value. This is an attempt to bypass the need to initiate notificationData with a variable of type T, ScrollNotification in this instance.

The previous commit stored doubles in NotificationNotifier instead of ScrollNotifications in this attempt. Not sure that this approach will ever work.